research-news-app/functions/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.1.0
google-cloud-secret-manager>=2.16.3
google-cloud-aiplatform>=1.38.1
PyPDF2>=3.0.1
python-magic>=0.4.27
Pillow>=10.0.1
langdetect>=1.0.9